feat: add .gitignore file to exclude unnecessary files from version control
This commit adds a .gitignore file to the project. This file is used to specify which files and directories should be ignored by Git, meaning they won't be included in any commits. This is useful for excluding files that are generated during the build process, log files, local configuration files, and any other files that don't need to be version controlled.